♻️ Adjust the import paths in the implementations for the refactoring.
