+++ test/unit_test/agent_core/test_router_and_registry.py
@@ -9,19 +9,19 @@ from gearmeshing_ai.agent_core.agent_registry import AgentRegistry
 from gearmeshing_ai.agent_core.factory import build_agent_registry
 from gearmeshing_ai.agent_core.planning.planner import StructuredPlanner
 from gearmeshing_ai.agent_core.policy.models import PolicyConfig
-from gearmeshing_ai.agent_core.role_provider import (
+from gearmeshing_ai.info_provider import (
     CognitiveProfile,
     RoleDefinition,
     RolePermissions,
     StaticAgentRoleProvider,
+    AgentRole,
+    CapabilityName,
 )
 from gearmeshing_ai.agent_core.router import Router
 from gearmeshing_ai.agent_core.runtime.models import EngineDeps
 from gearmeshing_ai.agent_core.schemas.domain import (
-    AgentRole,
     AgentRun,
     AutonomyProfile,
-    CapabilityName,
 )
 from gearmeshing_ai.agent_core.service import AgentServiceDeps
